Run the concurrency tests in an existing CI job
Each job in the test matrix uses the Vuforia credentials file which
matches its index, and there are exactly 100 of those files, so adding a
job made the last one fail to start.

Move the concurrency tests into ``test_flask_app_usage.py``, which
already has a job, and let that module's mock pass requests to a local
port through to a real server.
