+++ .github/workflows/test.yml
@@ -118,7 +118,6 @@ jobs:
           - tests/mock_vws/test_requests_mock_usage.py
           - tests/mock_vws/test_respx_mock_usage.py
           - tests/mock_vws/test_flask_app_usage.py
-          - tests/mock_vws/test_flask_app_concurrency.py
           - tests/mock_vws/test_model_target_generation_failure.py
           - tests/mock_vws/test_model_target_generation_warning.py
           - tests/mock_vws/test_model_target_web_api.py
